feat: invert control flow (#62)
This changes the implementation of the runner to allow for
better control of long-running steps.

Control is inverted and given to the step runners so they
can implement the best strategy for retrying.

Support for retry, tags, and configuration through tags
is removed. Step runners need to take care to ensure that
described asserting is met, without retry, because retrying
scenarios creates mostly too much noise and makes
it hard to determine the root cause. It also makes
flaky "normal", which should not be the case.

BREAKING CHANGE: this changes the API

See https://github.com/NordicSemiconductor/cloud-bdd-markdown-e2e-example-aws-js/pull/243 for a migration example